Beginnings of VM subsystem

+++ src/kern/arch/riscv/include/memlayout.h
@@ -1,4 +1,5 @@
 // Memory layout
+#include "types.h"
 
 #define EXTMEM  0x100000            // Start of extended memory
 #define PHYSTOP 0xE000000           // Top physical memory
@@ -10,12 +11,12 @@
 
 #ifndef __ASSEMBLER__
 
-static inline uint v2p(void *a) { return ((uint) (a))  - KERNBASE; }
-static inline void *p2v(uint a) { return (void *) ((a) + KERNBASE); }
+static inline ulong v2p(void *a) { return ((ulong) (a))  - KERNBASE; }
+static inline void *p2v(ulong a) { return (void *) ((a) + KERNBASE); }
 
 #endif
 
-#define V2P(a) (((uint) (a)) - KERNBASE)
+#define V2P(a) (((ulong) (a)) - KERNBASE)
 #define P2V(a) (((void *) (a)) + KERNBASE)
 
 #define V2P_WO(x) ((x) - KERNBASE)    // same as V2P, but without casts
